Clean up pass
- Moved RPC class into its own file.
