+++ amqpstorm/io.py
@@ -52,6 +52,7 @@ class Poller(object):
 
 class IO(object):
     """AMQP Connection.io"""
+
     def __init__(self, parameters, on_read=None):
         self._inbound_thread = None
         self._lock = threading.Lock()
