+++ amqpstorm/tests/queue_tests.py
@@ -1,5 +1,4 @@
 import logging
-import uuid
 
 try:
     import unittest2 as unittest
